+++ contracts/nois-drand/src/contract.rs
@@ -1,3 +1,5 @@
+#![allow(deprecated)]
+
 use cosmwasm_std::{
     ensure_eq, entry_point, to_binary, Attribute, BankMsg, Coin, CosmosMsg, Deps, DepsMut, Empty,
     Env, HexBinary, MessageInfo, Order, QueryResponse, Response, StdError, StdResult, Uint128,
@@ -13,8 +15,8 @@ use crate::attributes::{
 use crate::bots::{eligible_group, group, validate_moniker};
 use crate::error::ContractError;
 use crate::msg::{
-    AllowListResponse, BeaconResponse, BeaconsResponse, BotResponse, BotsResponse, ConfigResponse,
-    ExecuteMsg, InstantiateMsg, IsAllowListedResponse, NoisGatewayExecuteMsg, QueriedSubmission,
+    AllowlistResponse, BeaconResponse, BeaconsResponse, BotResponse, BotsResponse, ConfigResponse,
+    ExecuteMsg, InstantiateMsg, IsAllowlistedResponse, NoisGatewayExecuteMsg, QueriedSubmission,
     QueryMsg, SubmissionsResponse,
 };
 use crate::state::{
@@ -107,8 +109,10 @@ pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> StdResult<QueryResponse> {
         QueryMsg::Submissions { round } => to_binary(&query_submissions(deps, round)?)?,
         QueryMsg::Bot { address } => to_binary(&query_bot(deps, address)?)?,
         QueryMsg::Bots {} => to_binary(&query_bots(deps)?)?,
-        QueryMsg::AllowList {} => to_binary(&query_allow_list(deps)?)?,
-        QueryMsg::IsAllowListed { bot } => to_binary(&query_is_allow_listed(deps, bot)?)?,
+        QueryMsg::AllowList {} => to_binary(&query_allowlist(deps)?)?,
+        QueryMsg::Allowlist {} => to_binary(&query_allowlist(deps)?)?,
+        QueryMsg::IsAllowListed { bot } => to_binary(&query_is_allowlisted(deps, bot)?)?,
+        QueryMsg::IsAllowlisted { bot } => to_binary(&query_is_allowlisted(deps, bot)?)?,
     };
     Ok(response)
 }
@@ -182,7 +186,7 @@ fn query_bots(deps: Deps) -> StdResult<BotsResponse> {
     Ok(BotsResponse { bots })
 }
 
-fn query_allow_list(deps: Deps) -> StdResult<AllowListResponse> {
+fn query_allowlist(deps: Deps) -> StdResult<AllowlistResponse> {
     // No pagination here yet 🤷‍♂️
     let allowed = ALLOWLIST
         .range(deps.storage, None, None, Order::Ascending)
@@ -191,13 +195,13 @@ fn query_allow_list(deps: Deps) -> StdResult<AllowListResponse> {
             address.into()
         })
         .collect();
-    Ok(AllowListResponse { allowed })
+    Ok(AllowlistResponse { allowed })
 }
 
-fn query_is_allow_listed(deps: Deps, bot: String) -> StdResult<IsAllowListedResponse> {
+fn query_is_allowlisted(deps: Deps, bot: String) -> StdResult<IsAllowlistedResponse> {
     let bot_addr = deps.api.addr_validate(&bot)?;
     let listed = ALLOWLIST.has(deps.storage, &bot_addr);
-    Ok(IsAllowListedResponse { listed })
+    Ok(IsAllowlistedResponse { listed })
 }
 
 fn execute_register_bot(
@@ -713,7 +717,7 @@ mod tests {
         };
         instantiate(deps.as_mut(), mock_env(), mock_info("creator", &[]), msg).unwrap();
 
-        let IsAllowListedResponse { listed } = from_binary(
+        let IsAllowlistedResponse { listed } = from_binary(
             &query(
                 deps.as_ref(),
                 mock_env(),
@@ -742,7 +746,7 @@ mod tests {
         let info = mock_info(TESTING_MANAGER, &[]);
         execute(deps.as_mut(), mock_env(), info, msg).unwrap();
 
-        let IsAllowListedResponse { listed } = from_binary(
+        let IsAllowlistedResponse { listed } = from_binary(
             &query(
                 deps.as_ref(),
                 mock_env(),
@@ -763,7 +767,7 @@ mod tests {
         let info = mock_info(TESTING_MANAGER, &[]);
         execute(deps.as_mut(), mock_env(), info, msg).unwrap();
 
-        let IsAllowListedResponse { listed } = from_binary(
+        let IsAllowlistedResponse { listed } = from_binary(
             &query(
                 deps.as_ref(),
                 mock_env(),
@@ -784,7 +788,7 @@ mod tests {
         let info = mock_info(TESTING_MANAGER, &[]);
         execute(deps.as_mut(), mock_env(), info, msg).unwrap();
 
-        let IsAllowListedResponse { listed } = from_binary(
+        let IsAllowlistedResponse { listed } = from_binary(
             &query(
                 deps.as_ref(),
                 mock_env(),
@@ -1602,7 +1606,7 @@ mod tests {
     }
 
     #[test]
-    fn query_allow_list_works() {
+    fn query_allowlist_works() {
         let mut deps = mock_dependencies();
 
         let info = mock_info("creator", &[]);
@@ -1614,7 +1618,7 @@ mod tests {
         };
         instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
 
-        let AllowListResponse { allowed } =
+        let AllowlistResponse { allowed } =
             from_binary(&query(deps.as_ref(), mock_env(), QueryMsg::AllowList {}).unwrap())
                 .unwrap();
         assert_eq!(allowed, Vec::<String>::new());
@@ -1632,7 +1636,7 @@ mod tests {
         )
         .unwrap();
 
-        let AllowListResponse { allowed } =
+        let AllowlistResponse { allowed } =
             from_binary(&query(deps.as_ref(), mock_env(), QueryMsg::AllowList {}).unwrap())
                 .unwrap();
         assert_eq!(allowed, vec!["bot_b".to_string()]);
@@ -1650,7 +1654,7 @@ mod tests {
         )
         .unwrap();
 
-        let AllowListResponse { allowed } =
+        let AllowlistResponse { allowed } =
             from_binary(&query(deps.as_ref(), mock_env(), QueryMsg::AllowList {}).unwrap())
                 .unwrap();
         assert_eq!(
@@ -1791,7 +1795,7 @@ mod tests {
     }
 
     #[test]
-    fn is_query_allow_listed_works() {
+    fn is_query_allowlisted_works() {
         let mut deps = mock_dependencies();
 
         let info = mock_info("creator", &[]);
@@ -1817,7 +1821,7 @@ mod tests {
         .unwrap();
 
         // bot_b is listed
-        let IsAllowListedResponse { listed } = from_binary(
+        let IsAllowlistedResponse { listed } = from_binary(
             &query(
                 deps.as_ref(),
                 mock_env(),
@@ -1831,7 +1835,7 @@ mod tests {
         assert!(listed);
 
         // bot_a is not listed
-        let IsAllowListedResponse { listed } = from_binary(
+        let IsAllowlistedResponse { listed } = from_binary(
             &query(
                 deps.as_ref(),
                 mock_env(),