+++ contracts/nois-drand/src/msg.rs
@@ -1,3 +1,5 @@
+#![allow(deprecated)]
+
 use cosmwasm_schema::{cw_serde, QueryResponses};
 use cosmwasm_std::{Addr, HexBinary, Timestamp, Uint128};
 
@@ -74,11 +76,22 @@ pub enum QueryMsg {
     #[returns(BotsResponse)]
     Bots {},
     /// Returns the list of allowed bot addresses
-    #[returns(AllowListResponse)]
+    #[returns(AllowlistResponse)]
+    #[deprecated(note = "use QueryMsg::Allowlist")]
     AllowList {},
     /// Returns the list of allowed bot addresses
-    #[returns(IsAllowListedResponse)]
+    #[returns(AllowlistResponse)]
+    Allowlist {},
+    /// Returns if the given bot addresses is in the allowlist.
+    #[returns(IsAllowlistedResponse)]
+    #[deprecated(note = "use QueryMsg::IsAllowlisted")]
     IsAllowListed {
+        /// The address of the bot
+        bot: String,
+    },
+    /// Returns if the given bot addresses is in the allowlist.
+    #[returns(IsAllowlistedResponse)]
+    IsAllowlisted {
         /// The address of the bot
         bot: String,
     },
@@ -108,13 +121,13 @@ pub struct BotsResponse {
 }
 
 #[cw_serde]
-pub struct AllowListResponse {
+pub struct AllowlistResponse {
     /// List of bot addresses
     pub allowed: Vec<String>,
 }
 
 #[cw_serde]
-pub struct IsAllowListedResponse {
+pub struct IsAllowlistedResponse {
     pub listed: bool,
 }